functional partial cheep view done
