+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,10 +1,10 @@
 ﻿@page "/"
-@model Chirp.Web.Razor.PublicTimeLine
+@model Chirp.Web.models.PublicTimeLine
+@using Chirp.Web.models
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "./Shared/_Layout.cshtml";
 }
-
 
 <h2> Public Timeline </h2>
 
@@ -13,15 +13,11 @@
 
     @foreach (var cheep in Model.Cheeps)
     {
-        <div>
-            <div class="cheep-item">
-                <strong class="cheep-author">  <a href="/@cheep.Author">@cheep.Author</a> </strong>
-
-                <p class="cheep-body"> @cheep.Text </p>
-
-                <span class="cheep-date"> &mdash; @Model.ToDateTime(cheep.Timestamp) </span>
-            </div>
-        </div>
+        @await Html.PartialAsync("_Cheep", new CheepPartialView {
+            Author= cheep.Author,
+            Body= cheep.Text,
+            Date= cheep.Timestamp
+        })
     }
 
 } else
