+++ src/Chirp.Web/models/NavBarPartialView.cshtml.cs
@@ -1,0 +1,5 @@
+namespace Chirp.Web.models;
+
+public class NavBarPartialView {
+    
+}
